Move code_generator out in a dedicated file
